pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "model-monitor-analysis"
version = "0.1.0"
description = "BeeFrame model performance monitoring and analysis"
requires-python = ">=3.11"
dependencies = [
    "pandas",
    "pyyaml",
    "boto3",
    "psycopg2-binary",
    "sqlalchemy",
]

[project.optional-dependencies]
dev = [
    "pytest",
    "pytest-cov",
    "ruff",
]

[tool.hatch.build.targets.wheel]
packages = ["src/model_monitor"]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "--tb=short"

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "UP"]
